+++ NEWS.md
@@ -1,8 +1,10 @@
 # editData 0.1.2
 =================
-(2017-Oct-3)
+(2017-Oct-6)
 
 * Bug fixed
+
+* Add several example applications in "inst" folder
 
 # editData 0.1.1
 ==================